+++ src/Request/HandlerInterface.php
@@ -13,6 +13,7 @@ namespace Xabbuh\XApi\Client\Request;
 
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\StreamInterface;
 use Xabbuh\XApi\Common\Exception\XApiException;
 
 /**
@@ -27,12 +28,13 @@ interface HandlerInterface
      * @param string $uri           The URI to send the request to
      * @param array  $urlParameters Optional url parameters
      * @param string $body          An optional request body
+     * @param array  $headers       Optional additional HTTP headers
      *
      * @return RequestInterface The request
      *
      * @throws \InvalidArgumentException when no valid HTTP method is given
      */
-    public function createRequest($method, $uri, array $urlParameters = array(), $body = null);
+    public function createRequest($method, $uri, array $urlParameters = array(), $body = null, array $headers = array());
 
     /**
      * Performs the given HTTP request.